+++ LCLandmineOutside/LCHazardsOutside.cs
@@ -12,7 +12,7 @@ namespace LCHazardsOutside
     public class LCHazardsOutside : BaseUnityPlugin {
         private const string modGUID = "snake.tech.LCHazardsOutside";
         private const string modName = "LCHazardsOutside";
-        private const string modVersion = "1.0.0.0";
+        private const string modVersion = "1.0.1.0";
 
         private readonly Harmony harmony = new(modGUID);
 
@@ -45,7 +45,7 @@ namespace LCHazardsOutside
 
         void LoadConfig() {
             configEnableLandmine = Config.Bind("Landmine","EnableLandmineOutside", true, "Whether or not to spawn landmines outside."); 
-            configLandmineMinSpawnRate = Config.Bind("Landmine", "LandmineMinSpawnRate", 2, "Minimum amount of landmines to spawn outside.");
+            configLandmineMinSpawnRate = Config.Bind("Landmine", "LandmineMinSpawnRate", 5, "Minimum amount of landmines to spawn outside.");
             configLandmineMaxSpawnRate = Config.Bind("Landmine", "LandmineMaxSpawnRate", 15, "Maximum amount of landmines to spawn outside.");
 
             configEnableTurret = Config.Bind("Turret", "EnableTurretOutside", true, "Whether or not to spawn turrets outside.");
@@ -53,8 +53,8 @@ namespace LCHazardsOutside
             configTurretMaxSpawnRate = Config.Bind("Turret", "TurretMaxSpawnRate", 1, "Maximum amount of turrets to spawn outside.");
 
             configEnableCustomHazard = Config.Bind("Custom", "EnableCustomHazardOutside", true, "Whether or not to spawn modded hazards outside.");
-            configCustomHazardMinSpawnRate = Config.Bind("Custom", "CustomHazardMinSpawnRate", 2, "Minimum amount of custom hazards to spawn outside.");
-            configCustomHazardMaxSpawnRate = Config.Bind("Custom", "CustomHazardMaxSpawnRate", 5, "Maximum amount of custom hazards to spawn outside.");
+            configCustomHazardMinSpawnRate = Config.Bind("Custom", "CustomHazardMinSpawnRate", 1, "Minimum amount of custom hazards to spawn outside.");
+            configCustomHazardMaxSpawnRate = Config.Bind("Custom", "CustomHazardMaxSpawnRate", 2, "Maximum amount of custom hazards to spawn outside.");
         }
 
         public static ManualLogSource GetLogger() {