+++ LCLandmineOutside/Thunderstore/CHANGELOG.md
@@ -1,3 +1,9 @@
+v1.0.1
+======
+- Fixed hazards spawning in mid-air by improving the spawn method.
+- Increased default minimum spawn rate for landmines to 5.
+- Updated README.
+
 v1.0.0
 ======